Feat: 커뮤니티 게시글 상세 api로부터 받아오기
- tag, 이미지, 댓글 해야함

+++ app/src/main/java/com/project/veganlife/community/ui/view/CommunityHomeFragment.kt
@@ -167,7 +167,7 @@ class CommunityHomeFragment : Fragment(), CommunityFeedAdapter.OnItemClickListen
 
     override fun onItemClicked(item: PostPreview) {
         val bundle = Bundle().apply {
-            putLong("postId", item.id)
+            putInt("postId", item.id)
         }
         findNavController().navigate(R.id.action_communityHomeFragment_to_communityDetailFeedFragment, bundle)
     }
